+++ python/tests/mantid_scipp_comparison/test_histogram_events.py
@@ -12,9 +12,7 @@ class HistogramEventsTest(MantidScippComparison):
 
     @property
     def _filenames(self):
-        return {
-            "CNCS_51936_event.nxs": ("5ba401e489260a44374b5be12b780911", "MD5")
-        }
+        return {"CNCS_51936_event.nxs": ("5ba401e489260a44374b5be12b780911", "MD5")}
 
     def _run_mantid(self, input):
         import mantid.simpleapi as sapi
@@ -28,15 +26,12 @@ class HistogramEventsTest(MantidScippComparison):
     def _run_scipp(self, input):
         return sc.histogram(x=input,
                             bins=sc.Variable(dims=['tof'],
-                                             values=np.linspace(0,
-                                                                1000,
-                                                                num=101),
+                                             values=np.linspace(0, 1000, num=101),
                                              dtype=sc.dtype.float64,
                                              unit=sc.units.us))
 
 
-@pytest.mark.skipif(not mantid_is_available(),
-                    reason='Mantid framework is unavailable')
+@pytest.mark.skipif(not mantid_is_available(), reason='Mantid framework is unavailable')
 def test_histogram_events():
     test = HistogramEventsTest()
     test.run()